+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Removed `boundary=EJECT` from `Key` and `AvEnableFlags`
   for consistency with other `IntFlag`s.
+- Make `CompositeEnvironmentDriver` non-generic.
 
 ### Fixed